fix sqa wrapper and fix bugs
