+++ docs/apis/openjij.sampler.html
@@ -563,8 +563,8 @@ in which chimera unit cells are arranged.</p></li>
 </dd></dl>
 
 <dl class="method">
-<dt id="openjij.sampler.response.Response.update_quantum_ising_states_energies">
-<code class="descname">update_quantum_ising_states_energies</code><span class="sig-paren">(</span><em>trotter_states</em>, <em>q_energies</em><span class="sig-paren">)</span><a class="headerlink" href="#openjij.sampler.response.Response.update_quantum_ising_states_energies" title="Permalink to this definition">¶</a></dt>
+<dt id="openjij.sampler.response.Response.update_trotter_ising_states_energies">
+<code class="descname">update_trotter_ising_states_energies</code><span class="sig-paren">(</span><em>trotter_states</em>, <em>q_energies</em><span class="sig-paren">)</span><a class="headerlink" href="#openjij.sampler.response.Response.update_trotter_ising_states_energies" title="Permalink to this definition">¶</a></dt>
 <dd><p>Update quantum states and energies.</p>
 <dl class="field-list simple">
 <dt class="field-odd">Parameters</dt>
