+++ docs/genindex.html
@@ -422,7 +422,7 @@
 </li>
   </ul></td>
   <td style="width: 33%; vertical-align: top;"><ul>
-      <li><a href="apis/openjij.sampler.html#openjij.sampler.response.Response.update_quantum_ising_states_energies">update_quantum_ising_states_energies() (openjij.sampler.response.Response method)</a>
+      <li><a href="apis/openjij.sampler.html#openjij.sampler.response.Response.update_trotter_ising_states_energies">update_trotter_ising_states_energies() (openjij.sampler.response.Response method)</a>
 </li>
   </ul></td>
 </tr></table>
