+++ docs/searchindex.js
@@ -1,1 +1,1 @@
-Search.setIndex({docnames:["apis/modules","apis/openjij","apis/openjij.model","apis/openjij.sampler","index"],envversion:{"sphinx.domains.c":1,"sphinx.domains.changeset":1,"sphinx.domains.cpp":1,"sphinx.domains.javascript":1,"sphinx.domains.math":2,"sphinx.domains.python":1,"sphinx.domains.rst":1,"sphinx.domains.std":1,sphinx:56},filenames:["apis/modules.rst","apis/openjij.rst","apis/openjij.model.rst","apis/openjij.sampler.rst","index.rst"],objects:{"":{openjij:[1,0,0,"-"]},"openjij.model":{chimera_graph:[2,0,0,"-"],king_graph:[2,0,0,"-"],model:[2,0,0,"-"]},"openjij.model.chimera_graph":{ChimeraModel:[2,1,1,""]},"openjij.model.chimera_graph.ChimeraModel":{chimera_coordinate:[2,2,1,""],full_indices:[2,2,1,""],to_index:[2,2,1,""],validate_chimera:[2,2,1,""]},"openjij.model.king_graph":{KingGraph:[2,1,1,""]},"openjij.model.king_graph.KingGraph":{convert_ising:[2,2,1,""],convert_to_index:[2,2,1,""],get_ising_king_graph:[2,2,1,""],king_graph:[2,3,1,""],king_indices:[2,2,1,""],prange:[2,3,1,""],yrange:[2,3,1,""]},"openjij.model.model":{BinaryQuadraticModel:[2,1,1,""]},"openjij.model.model.BinaryQuadraticModel":{calc_energy:[2,2,1,""],convert_to_dense_graph:[2,2,1,""],interactions:[2,2,1,""],ising_dictionary:[2,2,1,""],ising_interactions:[2,2,1,""]},"openjij.sampler":{cmos_annealer:[3,0,0,"-"],gpu_sampler:[3,0,0,"-"],response:[3,0,0,"-"],sampler:[3,0,0,"-"]},"openjij.sampler.cmos_annealer":{CMOSAnnealer:[3,1,1,""]},"openjij.sampler.cmos_annealer.CMOSAnnealer":{cmos_parameters:[3,3,1,""],make_json_request:[3,2,1,""],sample_ising:[3,2,1,""],sample_qubo:[3,2,1,""]},"openjij.sampler.gpu_sampler":{GPUSQASampler:[3,1,1,""]},"openjij.sampler.gpu_sampler.GPUSQASampler":{energy_bias:[3,3,1,""],indices:[3,3,1,""],sample_ising:[3,2,1,""],sample_qubo:[3,2,1,""],sampling:[3,2,1,""],sqa_kwargs:[3,3,1,""],var_type:[3,3,1,""]},"openjij.sampler.response":{Response:[3,1,1,""]},"openjij.sampler.response.Response":{energies:[3,3,1,""],info:[3,3,1,""],min_samples:[3,3,1,""],q_energies:[3,3,1,""],q_states:[3,3,1,""],samples:[3,3,1,""],states:[3,3,1,""],update_ising_states_energies:[3,2,1,""],update_quantum_ising_states_energies:[3,2,1,""]},"openjij.sampler.sampler":{BaseSampler:[3,1,1,""],SASampler:[3,1,1,""],SQASampler:[3,1,1,""],measure_time:[3,4,1,""]},"openjij.sampler.sampler.SASampler":{N:[3,3,1,""],energy_bias:[3,3,1,""],indices:[3,3,1,""],sample_ising:[3,2,1,""],sample_qubo:[3,2,1,""],var_type:[3,3,1,""]},"openjij.sampler.sampler.SQASampler":{N:[3,3,1,""],energy_bias:[3,3,1,""],indices:[3,3,1,""],sample_ising:[3,2,1,""],sample_qubo:[3,2,1,""],sqa_kwargs:[3,3,1,""],var_type:[3,3,1,""]},"openjij.utils":{benchmark:[1,4,1,""],convert_response:[1,4,1,""]},openjij:{model:[2,0,0,"-"],sampler:[3,0,0,"-"],utils:[1,0,0,"-"]}},objnames:{"0":["py","module","Python module"],"1":["py","class","Python class"],"2":["py","method","Python method"],"3":["py","attribute","Python attribute"],"4":["py","function","Python function"]},objtypes:{"0":"py:module","1":"py:class","2":"py:method","3":"py:attribute","4":"py:function"},terms:{"1q_0q_2":4,"1q_0q_3":[],"2q_0":4,"2q_1":4,"8ly":2,"\u3053\u308c\u306fmarkdown\u3092\u4f7f\u3063\u305f\u30c9\u30ad\u30e5\u30e1\u30f3\u30c8\u3067\u3059":[],"\u3053\u3093\u306a\u611f\u3058\u3067\u3059":[],"\u30ea\u30b9\u30c8\u306f":[],"\u4f8b\u3048\u3070":[],"class":[2,3],"float":3,"import":[3,4],"int":[2,3],"markdown\u3092\u4f7f\u3063\u305f\u30c9\u30ad\u30e5\u30e1\u30f3\u30c8":[],"return":3,"switch":4,"try":4,For:4,Ising:3,One:2,The:4,Then:4,There:4,With:4,about:3,access:3,adjust:4,algorithm:4,all:3,amplitud:3,ani:[],anneal:[2,3,4],api:[2,3],arg:[],argument:[3,4],arrai:[3,4],arrang:3,asic:3,attribut:2,attributeerror:3,avail:4,averaged_energi:3,averaged_spin:3,base:[2,3],basesampl:3,below:3,benchmark:1,beta:3,beta_max:[3,4],beta_min:[3,4],beta_sch:4,bia:3,bias:3,binari:[3,4],binaryquadraticmodel:[2,3],bodi:[],bqm:2,calc_energi:2,can:4,carlo:3,cell:[2,3],chang:4,chimera:[2,3],chimera_coordin:2,chimera_graph:[0,1],chimeramodel:2,choos:4,chose:4,cloud:[2,3],cmo:[2,3],cmos_anneal:[0,1],cmos_paramet:3,cmosanneal:3,code:[],coeffici:[2,3],column:2,com:3,constant:[],constructor:3,content:0,control:4,convert:2,convert_is:2,convert_respons:1,convert_to_dense_graph:2,convert_to_index:2,cool:4,coordin:2,cpu:4,cxxjij:[2,3],decreas:4,dens:2,detail:3,develop:4,dict:3,dictionari:[2,4],dimens:2,dimension:3,doctest:3,doesn:3,each:[],easili:4,energi:[3,4],energy_bia:3,error:3,exampl:[3,4],execution_tim:[1,3],fals:2,featur:4,field:4,finish:4,fluctuat:3,format:2,fpga:3,from:[2,3],full_indic:2,func:3,gamma:3,geometr:4,geq:[],get:[3,4],get_ising_king_graph:2,github:[],given:3,gpu:3,gpu_sampl:[0,1],gpusasampl:3,gpusqasampl:3,graph:[2,3],ground_energi:1,handl:[],hardwar:[2,4],here:4,high:4,hitachi:2,how:4,html:3,http:3,incompat:3,increas:[],index:4,indic:3,info:3,inform:3,inherit:3,initi:[],input:3,instanti:4,integ:3,interact:[2,3],invers:[3,4],ising:3,ising_dictionari:2,ising_interact:2,iter:[1,3,4],job_id:3,keyword:3,king:[2,3],king_graph:[0,1,3],king_indic:2,kinggraph:2,kwarg:3,lattic:3,length:3,leq:4,less:3,let:4,linear:[2,3],list:[2,3],machine_typ:[2,3],make:3,make_json_request:3,mani:4,math:[],matrix:4,maximum:[3,4],mcmc:4,mean:3,measure_tim:3,method:[3,4],min_energi:4,min_sampl:[3,4],min_stat:4,minimum:3,minimun:3,minmum:[3,4],model:[0,1,3],modul:[0,4],mont:3,more:3,multipl:4,next:4,none:[2,3],num_occurr:4,number:[3,4],numpi:3,obj:3,object:[2,3],obtain:4,occurr:3,odd:3,one:3,onli:[3,4],optim:3,optin:[],option:[3,4],other:[3,4],out:4,p_d:1,packag:0,page:4,paramet:[3,4],perform:4,platform:4,pleas:[3,4],pmax:2,pmin:2,prang:2,print:3,problem:3,provid:4,python:[2,4],q_0:4,q_1:4,q_2:4,q_3:[],q_energi:3,q_i:4,q_j:4,q_state:3,quadrat:[2,3],quantum:[3,4],quantumis:3,qubo:3,rais:3,rang:3,re_calcul:2,reduc:[],refer:3,relat:2,repres:[2,3,4],represent:2,reprezent:2,request:3,research:4,respons:[0,1,4],restrict:2,result:[],revers:4,rubi:[],run:[],sampl:[3,4],sample_is:3,sample_qubo:[3,4],sampler:[0,1,4],sasampl:[3,4],schedul:[3,4],schedule_info:[3,4],search:4,see:4,show:[],side:3,simpl:4,simul:[3,4],skip:3,solut:4,solv:4,solver:1,some:[],specifi:[3,4],spin:[2,3],spin_typ:3,sqa:3,sqa_kwarg:3,sqasampl:[3,4],state:[2,3,4],step:3,step_length:3,step_num:3,str:3,strength:2,submit:3,submodul:0,subpackag:0,sum_:4,system:3,system_class:3,target:[],temperatur:[3,4],tempertur:[],term:2,than:3,therefor:4,thi:[3,4],time:4,time_nam:1,time_param_list:1,to_index:2,token:3,transvers:4,trotter:3,trotter_st:3,true_ground_st:1,tupl:[3,4],two:3,type:[2,3],unconstrain:3,unifi:[],unit:[2,3],unit_num_l:[2,3],updat:3,update_ising_states_energi:3,update_quantum_ising_states_energi:3,use:4,using:4,util:0,valid:3,validate_chimera:2,valu:[2,3],valueerror:3,var_typ:[2,3,4],variabl:3,vertic:3,violat:3,web:[2,3],when:4,which:[3,4],work:3,xmax:2,xmin:2,xrang:2,xxxxxxxxxxxxxxxxx:3,ymax:2,ymin:2,you:4,your_token:3,yrang:2,zero:3},titles:["openjij","openjij package","openjij.model package","openjij.sampler package","OpenJij \u2013 A unified framework for QUBO."],titleterms:{"\u4f8b\u3048\u3070":[],"markdown\u3092\u4f7f\u3063\u305f\u30c9\u30ad\u30e5\u30e1\u30f3\u30c8":[],Ising:[],chimera_graph:2,cmos_anneal:3,content:[1,2,3],document:[],framework:4,gpu_sampl:3,indic:4,king_graph:2,minimum:4,model:2,modul:[1,2,3],openjij:[0,1,2,3,4],packag:[1,2,3],qubo:4,respons:3,sampler:3,submodul:[1,2,3],subpackag:1,tabl:4,tutori:4,unifi:4,util:1,welcom:[]}})
+Search.setIndex({docnames:["apis/modules","apis/openjij","apis/openjij.model","apis/openjij.sampler","index"],envversion:{"sphinx.domains.c":1,"sphinx.domains.changeset":1,"sphinx.domains.cpp":1,"sphinx.domains.javascript":1,"sphinx.domains.math":2,"sphinx.domains.python":1,"sphinx.domains.rst":1,"sphinx.domains.std":1,sphinx:56},filenames:["apis/modules.rst","apis/openjij.rst","apis/openjij.model.rst","apis/openjij.sampler.rst","index.rst"],objects:{"":{openjij:[1,0,0,"-"]},"openjij.model":{chimera_graph:[2,0,0,"-"],king_graph:[2,0,0,"-"],model:[2,0,0,"-"]},"openjij.model.chimera_graph":{ChimeraModel:[2,1,1,""]},"openjij.model.chimera_graph.ChimeraModel":{chimera_coordinate:[2,2,1,""],full_indices:[2,2,1,""],to_index:[2,2,1,""],validate_chimera:[2,2,1,""]},"openjij.model.king_graph":{KingGraph:[2,1,1,""]},"openjij.model.king_graph.KingGraph":{convert_ising:[2,2,1,""],convert_to_index:[2,2,1,""],get_ising_king_graph:[2,2,1,""],king_graph:[2,3,1,""],king_indices:[2,2,1,""],prange:[2,3,1,""],yrange:[2,3,1,""]},"openjij.model.model":{BinaryQuadraticModel:[2,1,1,""]},"openjij.model.model.BinaryQuadraticModel":{calc_energy:[2,2,1,""],convert_to_dense_graph:[2,2,1,""],interactions:[2,2,1,""],ising_dictionary:[2,2,1,""],ising_interactions:[2,2,1,""]},"openjij.sampler":{cmos_annealer:[3,0,0,"-"],gpu_sampler:[3,0,0,"-"],response:[3,0,0,"-"],sampler:[3,0,0,"-"]},"openjij.sampler.cmos_annealer":{CMOSAnnealer:[3,1,1,""]},"openjij.sampler.cmos_annealer.CMOSAnnealer":{cmos_parameters:[3,3,1,""],make_json_request:[3,2,1,""],sample_ising:[3,2,1,""],sample_qubo:[3,2,1,""]},"openjij.sampler.gpu_sampler":{GPUSQASampler:[3,1,1,""]},"openjij.sampler.gpu_sampler.GPUSQASampler":{energy_bias:[3,3,1,""],indices:[3,3,1,""],sample_ising:[3,2,1,""],sample_qubo:[3,2,1,""],sampling:[3,2,1,""],sqa_kwargs:[3,3,1,""],var_type:[3,3,1,""]},"openjij.sampler.response":{Response:[3,1,1,""]},"openjij.sampler.response.Response":{energies:[3,3,1,""],info:[3,3,1,""],min_samples:[3,3,1,""],q_energies:[3,3,1,""],q_states:[3,3,1,""],samples:[3,3,1,""],states:[3,3,1,""],update_ising_states_energies:[3,2,1,""],update_trotter_ising_states_energies:[3,2,1,""]},"openjij.sampler.sampler":{BaseSampler:[3,1,1,""],SASampler:[3,1,1,""],SQASampler:[3,1,1,""],measure_time:[3,4,1,""]},"openjij.sampler.sampler.SASampler":{N:[3,3,1,""],energy_bias:[3,3,1,""],indices:[3,3,1,""],sample_ising:[3,2,1,""],sample_qubo:[3,2,1,""],var_type:[3,3,1,""]},"openjij.sampler.sampler.SQASampler":{N:[3,3,1,""],energy_bias:[3,3,1,""],indices:[3,3,1,""],sample_ising:[3,2,1,""],sample_qubo:[3,2,1,""],sqa_kwargs:[3,3,1,""],var_type:[3,3,1,""]},"openjij.utils":{benchmark:[1,4,1,""],convert_response:[1,4,1,""]},openjij:{model:[2,0,0,"-"],sampler:[3,0,0,"-"],utils:[1,0,0,"-"]}},objnames:{"0":["py","module","Python module"],"1":["py","class","Python class"],"2":["py","method","Python method"],"3":["py","attribute","Python attribute"],"4":["py","function","Python function"]},objtypes:{"0":"py:module","1":"py:class","2":"py:method","3":"py:attribute","4":"py:function"},terms:{"1q_0q_2":4,"1q_0q_3":[],"2q_0":4,"2q_1":4,"8ly":2,"\u3053\u308c\u306fmarkdown\u3092\u4f7f\u3063\u305f\u30c9\u30ad\u30e5\u30e1\u30f3\u30c8\u3067\u3059":[],"\u3053\u3093\u306a\u611f\u3058\u3067\u3059":[],"\u30ea\u30b9\u30c8\u306f":[],"\u4f8b\u3048\u3070":[],"class":[2,3],"float":3,"import":[3,4],"int":[2,3],"markdown\u3092\u4f7f\u3063\u305f\u30c9\u30ad\u30e5\u30e1\u30f3\u30c8":[],"return":3,"switch":4,"try":4,For:4,Ising:3,One:2,The:4,Then:4,There:4,With:4,about:3,access:3,adjust:4,algorithm:4,all:3,amplitud:3,ani:[],anneal:[2,3,4],api:[2,3],arg:[],argument:[3,4],arrai:[3,4],arrang:3,asic:3,attribut:2,attributeerror:3,avail:4,averaged_energi:3,averaged_spin:3,base:[2,3],basesampl:3,below:3,benchmark:1,beta:3,beta_max:[3,4],beta_min:[3,4],beta_sch:4,bia:3,bias:3,binari:[3,4],binaryquadraticmodel:[2,3],bodi:[],bqm:2,calc_energi:2,can:4,carlo:3,cell:[2,3],chang:4,chimera:[2,3],chimera_coordin:2,chimera_graph:[0,1],chimeramodel:2,choos:4,chose:4,cloud:[2,3],cmo:[2,3],cmos_anneal:[0,1],cmos_paramet:3,cmosanneal:3,code:[],coeffici:[2,3],column:2,com:3,constant:[],constructor:3,content:0,control:4,convert:2,convert_is:2,convert_respons:1,convert_to_dense_graph:2,convert_to_index:2,cool:4,coordin:2,cpu:4,cxxjij:[2,3],decreas:4,dens:2,detail:3,develop:4,dict:3,dictionari:[2,4],dimens:2,dimension:3,doctest:3,doesn:3,each:[],easili:4,energi:[3,4],energy_bia:3,error:3,exampl:[3,4],execution_tim:[1,3],fals:2,featur:4,field:4,finish:4,fluctuat:3,format:2,fpga:3,from:[2,3],full_indic:2,func:3,gamma:3,geometr:4,geq:[],get:[3,4],get_ising_king_graph:2,github:[],given:3,gpu:3,gpu_sampl:[0,1],gpusasampl:3,gpusqasampl:3,graph:[2,3],ground_energi:1,handl:[],hardwar:[2,4],here:4,high:4,hitachi:2,how:4,html:3,http:3,incompat:3,increas:[],index:4,indic:3,info:3,inform:3,inherit:3,initi:[],input:3,instanti:4,integ:3,interact:[2,3],invers:[3,4],ising:3,ising_dictionari:2,ising_interact:2,iter:[1,3,4],job_id:3,keyword:3,king:[2,3],king_graph:[0,1,3],king_indic:2,kinggraph:2,kwarg:3,lattic:3,length:3,leq:4,less:3,let:4,linear:[2,3],list:[2,3],machine_typ:[2,3],make:3,make_json_request:3,mani:4,math:[],matrix:4,maximum:[3,4],mcmc:4,mean:3,measure_tim:3,method:[3,4],min_energi:4,min_sampl:[3,4],min_stat:4,minimum:3,minimun:3,minmum:[3,4],model:[0,1,3],modul:[0,4],mont:3,more:3,multipl:4,next:4,none:[2,3],num_occurr:4,number:[3,4],numpi:3,obj:3,object:[2,3],obtain:4,occurr:3,odd:3,one:3,onli:[3,4],optim:3,optin:[],option:[3,4],other:[3,4],out:4,p_d:1,packag:0,page:4,paramet:[3,4],perform:4,platform:4,pleas:[3,4],pmax:2,pmin:2,prang:2,print:3,problem:3,provid:4,python:[2,4],q_0:4,q_1:4,q_2:4,q_3:[],q_energi:3,q_i:4,q_j:4,q_state:3,quadrat:[2,3],quantum:[3,4],quantumis:3,qubo:3,rais:3,rang:3,re_calcul:2,reduc:[],refer:3,relat:2,repres:[2,3,4],represent:2,reprezent:2,request:3,research:4,respons:[0,1,4],restrict:2,result:[],revers:4,rubi:[],run:[],sampl:[3,4],sample_is:3,sample_qubo:[3,4],sampler:[0,1,4],sasampl:[3,4],schedul:[3,4],schedule_info:[3,4],search:4,see:4,show:[],side:3,simpl:4,simul:[3,4],skip:3,solut:4,solv:4,solver:1,some:[],specifi:[3,4],spin:[2,3],spin_typ:3,sqa:3,sqa_kwarg:3,sqasampl:[3,4],state:[2,3,4],step:3,step_length:3,step_num:3,str:3,strength:2,submit:3,submodul:0,subpackag:0,sum_:4,system:3,system_class:3,target:[],temperatur:[3,4],tempertur:[],term:2,than:3,therefor:4,thi:[3,4],time:4,time_nam:1,time_param_list:1,to_index:2,token:3,transvers:4,trotter:3,trotter_st:3,true_ground_st:1,tupl:[3,4],two:3,type:[2,3],unconstrain:3,unifi:[],unit:[2,3],unit_num_l:[2,3],updat:3,update_ising_states_energi:3,update_trotter_ising_states_energi:3,use:4,using:4,util:0,valid:3,validate_chimera:2,valu:[2,3],valueerror:3,var_typ:[2,3,4],variabl:3,vertic:3,violat:3,web:[2,3],when:4,which:[3,4],work:3,xmax:2,xmin:2,xrang:2,xxxxxxxxxxxxxxxxx:3,ymax:2,ymin:2,you:4,your_token:3,yrang:2,zero:3},titles:["openjij","openjij package","openjij.model package","openjij.sampler package","OpenJij \u2013 A unified framework for QUBO."],titleterms:{"\u4f8b\u3048\u3070":[],"markdown\u3092\u4f7f\u3063\u305f\u30c9\u30ad\u30e5\u30e1\u30f3\u30c8":[],Ising:[],chimera_graph:2,cmos_anneal:3,content:[1,2,3],document:[],framework:4,gpu_sampl:3,indic:4,king_graph:2,minimum:4,model:2,modul:[1,2,3],openjij:[0,1,2,3,4],packag:[1,2,3],qubo:4,respons:3,sampler:3,submodul:[1,2,3],subpackag:1,tabl:4,tutori:4,unifi:4,util:1,welcom:[]}})
